refactor(af): rename to sangfor_af_v8_0_xx with conventional naming
Apply Sangfor naming convention used by sangfor_sip / sangfor_xdr:
- Plugin directory: af_v8_0_xx → sangfor_af_v8_0_xx
- Handler file:     af.handler.py → sangfor_af.handler.py
- Tool YAML files:  af_xxx.yaml → sangfor_af_v{48,85,106}_xxx.yaml
- Tool name field:  af_xxx → sangfor_af_v{48,85,106}_xxx
  (versioned tool names prevent registry collisions across the three
  versions, since the registry keys tools by name globally)

Provider/storage_key model now follows sangfor_sip/onesec:
- _provider.yaml service_id = sangfor_af (shared, unversioned)
- _provider.yaml version    = 8.0.48 / 8.0.85 / 8.0.106
- Flocks derives storage_key sangfor_af_v8_0_xx automatically so the
  three versions store credentials side-by-side under distinct keys.

Handler SERVICE_ID resolves to the full storage_key so each version
reads its own credentials.

Provider config cleanup:
- Drop redundant verify_ssl from credential_fields (WebUI renders the
  SSL toggle natively); add custom_settings.verify_ssl fallback path
  in _resolve_verify_ssl, matching sangfor_sip behavior.

Verified: all 21 tools (6+7+8) load without conflict via yaml_to_tool.
