--- .flocks/plugins/tools/api/af_v8_0_106/_provider.yaml
@@ -1,56 +1,0 @@
-name: af
-service_id: af_api_v8_0_106
-version: "8.0.106"
-description: >
-  Sangfor AF (Application Firewall) v8.0.106 REST API service (latest).
-  Includes all features from v8.0.85 plus extended session management
-  (recheck, export, block), alarm notifications, CPU/memory trend curves,
-  and link probe (SLA) APIs.
-description_cn: >
-  深信服 AF 下一代防火墙 v8.0.106 REST API 服务（最新版本）。
-  在 v8.0.85 基础上新增：会话精细化管理（recheck/导出/阻断）、
-  告警通知（在册事件、服务商信息、信息模板）、CPU/内存曲线、
-  链路探测（SLA）等接口。
-auth:
-  type: custom
-  secret: af_v8_0_106_username
-  secret_secret: af_v8_0_106_password
-credential_fields:
-  - key: username
-    label: 管理员用户名
-    storage: secret
-    config_key: username
-    secret_id: af_v8_0_106_username
-    input_type: text
-    required: true
-  - key: password
-    label: 管理员密码
-    storage: secret
-    config_key: password
-    secret_id: af_v8_0_106_password
-    input_type: password
-    required: true
-  - key: base_url
-    label: 设备地址 (Base URL)
-    storage: config
-    config_key: base_url
-    input_type: url
-    default: "https://192.168.1.1"
-  - key: verify_ssl
-    label: 验证 SSL 证书
-    storage: config
-    config_key: verify_ssl
-    input_type: text
-    default: "false"
-defaults:
-  base_url: "https://192.168.1.1"
-  timeout: 60
-  category: custom
-  product_version: "8.0.106"
-notes: |
-  深信服 AF v8.0.106 API 认证流程（同 v8.0.85）：
-  1. POST /api/v1/namespaces/public/login，传入用户名/密码，获取 token。
-  2. 后续请求在 HTTP 请求头中携带 Cookie: token=<token>。
-  3. token 默认 10 分钟无操作后失效，可通过 keepalive 接口刷新。
-  4. verify_ssl 默认 false（设备通常使用自签名证书）。
-  5. 新增：会话导出、会话阻断、告警通知、CPU/内存趋势曲线等 API。
