--- .flocks/plugins/tools/api/af_v8_0_106/af_ops.yaml
@@ -1,165 +1,0 @@
-name: af_ops
-description: >
-  Sangfor AF v8.0.48 operations center tool. Manages blacklist/whitelist
-  entries (IPs, domains, URLs) and blocked attacker IPs via the `action`
-  parameter. Key security triage actions for SOC workflows.
-description_cn: >
-  深信服 AF v8.0.48 运营中心工具。通过 `action` 参数管理黑白名单（IP/域名/URL）
-  和封锁攻击者 IP。是 SOC 安全处置的核心接口。
-category: custom
-enabled: true
-requires_confirmation: true
-provider: af_api_v8_0_106
-inputSchema:
-  type: object
-  properties:
-    action:
-      type: string
-      description: |
-        运营中心动作名，可选值：
-
-        ## 黑白名单管理
-        - get_blackwhitelist
-          用途: 查询黑白名单列表（IP/域名/URL）
-          必填: 无
-          常用: type（BLACK/WHITE）、_start、_length
-          风险提示: 只读查询接口
-          是否任务型: 否
-        - add_blackwhitelist
-          用途: 添加单条黑白名单
-          必填: url（IP/域名/URL）、type（BLACK/WHITE）
-          常用: enable、description、domain（0=IP,1=域名,2=URL）
-          风险提示: 写操作；添加黑名单会拦截对应流量
-          是否任务型: 否
-        - batch_add_blackwhitelist
-          用途: 批量添加黑白名单
-          必填: items（数组，每项含 url/type 字段）
-          风险提示: 写操作，批量添加黑名单影响面大
-          是否任务型: 否
-        - delete_blackwhitelist
-          用途: 删除单条黑白名单
-          必填: url（条目的 IP/域名/URL）
-          常用: type（BLACK/WHITE）
-          风险提示: 写操作，删除白名单可能导致误拦截
-          是否任务型: 否
-        - batch_delete_blackwhitelist
-          用途: 批量删除黑白名单
-          必填: items（数组，每项含 url 字段）
-          风险提示: 写操作，批量删除影响面大
-          是否任务型: 否
-
-        ## 封锁攻击者 IP
-        - get_blockip_list
-          用途: 查询当前封锁攻击者 IP 列表
-          必填: 无
-          常用: _start、_length、fuzzyIP（模糊搜索）、creator（AF/SIP）
-          风险提示: 只读查询接口
-          是否任务型: 否
-        - batch_add_blockip
-          用途: 批量封锁攻击者 IP
-          必填: items（数组，每项含 srcIP、dstIP 等字段）
-          常用: aifwType（MANUAL/AUTO）
-          风险提示: 高风险写操作；封锁 IP 会拦截其所有流量
-          是否任务型: 否
-        - batch_delete_blockip
-          用途: 批量解封攻击者 IP
-          必填: items（数组，每项含 srcIP、dstIP 等字段）
-          风险提示: 写操作，解封恶意 IP 存在安全风险
-          是否任务型: 否
-        - clear_blockip
-          用途: 清空封锁攻击者 IP 列表
-          必填: 无
-          常用: creator（AF/SIP，指定清除哪类封锁）
-          风险提示: 高风险写操作；会清除所有封锁 IP
-          是否任务型: 否
-        - get_blockip_auto_config
-          用途: 获取自动封锁攻击者时长配置
-          必填: 无
-          风险提示: 只读查询接口
-          是否任务型: 否
-        - set_blockip_auto_config
-          用途: 修改自动封锁攻击者时长
-          必填: blockTime（封锁时长，单位秒）
-          风险提示: 写操作，影响自动封锁策略
-          是否任务型: 否
-      enum:
-        - get_blackwhitelist
-        - add_blackwhitelist
-        - batch_add_blackwhitelist
-        - delete_blackwhitelist
-        - batch_delete_blackwhitelist
-        - get_blockip_list
-        - batch_add_blockip
-        - batch_delete_blockip
-        - clear_blockip
-        - get_blockip_auto_config
-        - set_blockip_auto_config
-
-    # Blacklist/whitelist params
-    url:
-      type: string
-      description: IP地址、域名或URL（黑白名单条目值）
-    type:
-      type: string
-      description: "名单类型：BLACK（黑名单）或 WHITE（白名单）"
-      enum:
-        - BLACK
-        - WHITE
-    enable:
-      type: boolean
-      description: 是否启用该条目，默认 true
-    description:
-      type: string
-      description: 条目描述信息（最大95字符）
-    domain:
-      type: integer
-      description: "条目类型：0=IP地址，1=域名，2=URL"
-      enum: [0, 1, 2]
-    items:
-      type: array
-      items:
-        type: object
-      description: 批量操作时的条目数组，每项至少包含 url（黑白名单）或 srcIP/dstIP（封锁IP）
-
-    # Block IP params
-    fuzzyIP:
-      type: string
-      description: 模糊搜索IP关键字（最大15字符）
-    creator:
-      type: string
-      description: "封锁来源身份：AF（防火墙自身）或 SIP（安全感知平台）"
-      enum:
-        - AF
-        - SIP
-    aifwType:
-      type: string
-      description: "添加封锁IP的类型：MANUAL（手动）或 AUTO（自动，需要 creator=SIP）"
-      enum:
-        - MANUAL
-        - AUTO
-    blockTime:
-      type: integer
-      description: 自动封锁时长（秒）
-
-    # Pagination
-    _start:
-      type: integer
-      description: 分页起始位置（从0开始）
-    _length:
-      type: integer
-      description: 每页最大返回数量（最大200，默认100）
-    _sortby:
-      type: string
-      description: 排序字段名
-    _order:
-      type: string
-      description: "排序方向：asc（升序）或 desc（降序）"
-      enum:
-        - asc
-        - desc
-  required:
-    - action
-handler:
-  type: script
-  script_file: af.handler.py
-  function: ops
