--- .flocks/plugins/tools/api/af_v8_0_106/af_status.yaml
@@ -1,120 +1,0 @@
-name: af_status
-description: >
-  Sangfor AF v8.0.48 device status tool. Query system resource usage
-  (CPU, memory, disk), firmware version, network interface status,
-  current time, and system uptime.
-description_cn: >
-  深信服 AF v8.0.48 状态中心工具。通过 `action` 参数查询系统资源（CPU/内存/磁盘）、
-  固件版本、网口状态、当前时间及系统运行时长等信息。
-category: custom
-enabled: true
-requires_confirmation: false
-provider: af_api_v8_0_106
-inputSchema:
-  type: object
-  properties:
-    action:
-      type: string
-      description: |
-        状态查询动作名，可选值：
-        - get_memory_usage
-          用途: 获取当前内存使用率（百分比）
-          必填: 无
-          风险提示: 只读查询接口
-          是否任务型: 否
-        - get_cpu_usage
-          用途: 获取当前 CPU 使用率（百分比）
-          必填: 无
-          风险提示: 只读查询接口
-          是否任务型: 否
-        - get_disk_usage
-          用途: 获取当前磁盘使用率（百分比）
-          必填: 无
-          风险提示: 只读查询接口
-          是否任务型: 否
-        - get_system_version
-          用途: 获取 AF 系统固件版本信息
-          必填: 无
-          常用: filter（ALL/FULL/MAJOR/MINOR 等）
-          风险提示: 只读查询接口
-          是否任务型: 否
-        - get_interface_status
-          用途: 获取指定网口或全部网口的状态（流速、连接状态）
-          必填: 无
-          常用: interfaceNames（如 eth0，不传则获取全部）
-          风险提示: 只读查询接口
-          是否任务型: 否
-        - get_runtime_status
-          用途: 获取系统运行时长（uptime）
-          必填: 无
-          风险提示: 只读查询接口
-          是否任务型: 否
-        - get_current_time
-          用途: 获取设备当前时间
-          必填: 无
-          风险提示: 只读查询接口
-          是否任务型: 否
-
-        ## 趋势数据（v8.0.106 新增）
-        - get_memory_trend
-          用途: 获取内存使用率变化曲线数据
-          必填: 无
-          常用: minutes（最近N分钟，默认60）
-          风险提示: 只读查询接口
-          是否任务型: 否
-        - get_cpu_trend
-          用途: 获取 CPU 使用率变化曲线数据
-          必填: 无
-          常用: minutes（最近N分钟，默认60）
-          风险提示: 只读查询接口
-          是否任务型: 否
-        - get_interface_throughput
-          用途: 获取指定接口的吞吐率折线图数据
-          必填: 无
-          常用: ifname（接口名，如 eth0）、minutes
-          风险提示: 只读查询接口
-          是否任务型: 否
-      enum:
-        - get_memory_usage
-        - get_cpu_usage
-        - get_disk_usage
-        - get_system_version
-        - get_interface_status
-        - get_runtime_status
-        - get_current_time
-        - get_memory_trend
-        - get_cpu_trend
-        - get_interface_throughput
-    filter:
-      type: string
-      description: >
-        版本信息过滤（仅用于 get_system_version）：
-        ALL=显示所有，FULL=完整版本号，MAJOR=主版本号，MINOR=次版本号，
-        INCREASE=增版本号，BUILD=创建日期，EN=是否英文版，HF=是否HF版，B=是否Beta版
-      enum:
-        - ALL
-        - FULL
-        - MAJOR
-        - MINOR
-        - INCREASE
-        - BUILD
-        - EN
-        - HF
-        - B
-        - R
-        - ADD
-    interfaceNames:
-      type: string
-      description: 网口名称（如 eth0），用于 get_interface_status；不填则获取全部接口
-    ifname:
-      type: string
-      description: 接口名称（如 eth0），用于 get_interface_throughput
-    minutes:
-      type: integer
-      description: 获取最近N分钟的趋势数据（用于 get_memory_trend/get_cpu_trend/get_interface_throughput）
-  required:
-    - action
-handler:
-  type: script
-  script_file: af.handler.py
-  function: status
