--- .flocks/plugins/tools/api/af_v8_0_106/af_system.yaml
@@ -1,53 +1,0 @@
-name: af_system
-description: >
-  Sangfor AF v8.0.48 system management tool. Query and manage administrator
-  accounts on the AF device.
-description_cn: >
-  深信服 AF v8.0.48 系统管理工具。通过 `action` 参数查询和管理 AF 设备上的
-  管理员账户信息。
-category: custom
-enabled: true
-requires_confirmation: true
-provider: af_api_v8_0_106
-inputSchema:
-  type: object
-  properties:
-    action:
-      type: string
-      description: |
-        系统管理动作名，可选值：
-        - get_accounts
-          用途: 查询所有管理员账户列表
-          必填: 无
-          常用: _start、_length、enable
-          风险提示: 只读查询接口
-          是否任务型: 否
-        - get_account
-          用途: 查询指定管理员账户详情
-          必填: name（账户名）
-          风险提示: 只读查询接口
-          是否任务型: 否
-      enum:
-        - get_accounts
-        - get_account
-    name:
-      type: string
-      description: 管理员账户名（用于 get_account）
-    enable:
-      type: boolean
-      description: 按启用/禁用状态过滤账户
-    _start:
-      type: integer
-      description: 分页起始位置（从0开始）
-    _length:
-      type: integer
-      description: 每页最大返回数量
-    _search:
-      type: string
-      description: 模糊搜索关键字
-  required:
-    - action
-handler:
-  type: script
-  script_file: af.handler.py
-  function: system
