--- .flocks/plugins/tools/api/af_v8_0_48/_provider.yaml
@@ -1,53 +1,0 @@
-name: af
-service_id: af_api_v8_0_48
-version: "8.0.48"
-description: >
-  Sangfor AF (Application Firewall) v8.0.48 REST API service.
-  Uses session-based cookie authentication: call login to obtain
-  a token, then supply it via Cookie header in subsequent requests.
-description_cn: >
-  深信服 AF 下一代防火墙 v8.0.48 REST API 服务。
-  使用基于 Session 的 Cookie 认证：首先调用 login 获取 token，
-  后续请求在 Cookie 中携带 token。基础 URL 为设备管理地址（如 https://192.168.1.1）。
-auth:
-  type: custom
-  secret: af_v8_0_48_username
-  secret_secret: af_v8_0_48_password
-credential_fields:
-  - key: username
-    label: 管理员用户名
-    storage: secret
-    config_key: username
-    secret_id: af_v8_0_48_username
-    input_type: text
-    required: true
-  - key: password
-    label: 管理员密码
-    storage: secret
-    config_key: password
-    secret_id: af_v8_0_48_password
-    input_type: password
-    required: true
-  - key: base_url
-    label: 设备地址 (Base URL)
-    storage: config
-    config_key: base_url
-    input_type: url
-    default: "https://192.168.1.1"
-  - key: verify_ssl
-    label: 验证 SSL 证书
-    storage: config
-    config_key: verify_ssl
-    input_type: text
-    default: "false"
-defaults:
-  base_url: "https://192.168.1.1"
-  timeout: 60
-  category: custom
-  product_version: "8.0.48"
-notes: |
-  深信服 AF v8.0.48 API 认证流程：
-  1. 调用 POST /api/v1/namespaces/public/login，传入用户名/密码，获取 token。
-  2. 后续所有请求在 HTTP 请求头中携带 Cookie: token=<token>。
-  3. token 默认 10 分钟无操作后失效，可通过 keepalive 接口刷新。
-  4. verify_ssl 默认 false（设备通常使用自签名证书）。
