--- .flocks/plugins/tools/api/af_v8_0_48/_test.yaml
@@ -1,103 +1,0 @@
-schema_version: 1
-provider: af_api_v8_0_48
-
-# Service-level connectivity probe.
-# get_system_version is a lightweight read-only GET that verifies
-# authentication and basic reachability.
-connectivity:
-  tool: af_status
-  params:
-    action: get_system_version
-
-# Tool-level test samples shown in the WebUI ToolDetailDrawer drop-down.
-fixtures:
-  af_auth:
-    - label: "Session keepalive"
-      label_cn: "刷新 Session 保活"
-      tags: [smoke]
-      params:
-        action: keepalive
-      assert:
-        success: true
-
-  af_status:
-    - label: "Get system version"
-      label_cn: "获取系统版本信息"
-      tags: [smoke]
-      params:
-        action: get_system_version
-      assert:
-        success: true
-
-    - label: "Get CPU usage"
-      label_cn: "获取 CPU 使用率"
-      tags: [smoke]
-      params:
-        action: get_cpu_usage
-
-    - label: "Get memory usage"
-      label_cn: "获取内存使用率"
-      tags: [smoke]
-      params:
-        action: get_memory_usage
-
-  af_ops:
-    - label: "List blacklist entries"
-      label_cn: "查询黑名单列表"
-      tags: [smoke, blacklist]
-      params:
-        action: get_blackwhitelist
-        type: BLACK
-      assert:
-        success: true
-
-    - label: "List whitelist entries"
-      label_cn: "查询白名单列表"
-      tags: [smoke, whitelist]
-      params:
-        action: get_blackwhitelist
-        type: WHITE
-
-    - label: "List blocked attacker IPs"
-      label_cn: "查询封锁攻击者 IP 列表"
-      tags: [smoke, blockip]
-      params:
-        action: get_blockip_list
-      assert:
-        success: true
-
-  af_objects:
-    - label: "List IP groups"
-      label_cn: "查询 IP 地址组列表"
-      tags: [smoke]
-      params:
-        action: get_ipgroups
-      assert:
-        success: true
-
-    - label: "List services"
-      label_cn: "查询服务对象列表"
-      tags: [smoke]
-      params:
-        action: get_services
-      assert:
-        success: true
-
-  af_network:
-    - label: "List routing table (all routes)"
-      label_cn: "查询完整路由表"
-      tags: [smoke]
-      params:
-        action: get_routes
-        routeType: ALL_ROUTE
-      assert:
-        success: true
-
-  af_system:
-    - label: "List admin accounts"
-      label_cn: "查询管理员账户列表"
-      tags: [smoke]
-      params:
-        action: get_accounts
-      assert:
-        success: true
