--- .flocks/plugins/tools/api/af_v8_0_48/af_objects.yaml
@@ -1,152 +1,0 @@
-name: af_objects
-description: >
-  Sangfor AF v8.0.48 objects management tool. Query, create, update, and
-  delete network IP group objects and services (protocol/port definitions)
-  used in firewall policies.
-description_cn: >
-  深信服 AF v8.0.48 对象管理工具。通过 `action` 参数查询、创建、修改和删除
-  IP 地址组对象及服务对象（协议/端口定义），这些对象被防火墙策略引用。
-category: custom
-enabled: true
-requires_confirmation: true
-provider: af_api_v8_0_48
-inputSchema:
-  type: object
-  properties:
-    action:
-      type: string
-      description: |
-        对象管理动作名，可选值：
-
-        ## IP 地址组
-        - get_ipgroups
-          用途: 查询符合条件的 IP 地址组列表
-          必填: 无
-          常用: _start、_length、businessType、__nameprefix、important、_search
-          风险提示: 只读查询接口
-          是否任务型: 否
-        - get_ipgroup
-          用途: 获取单个 IP 地址组详情
-          必填: uuid
-          风险提示: 只读查询接口
-          是否任务型: 否
-        - create_ipgroup
-          用途: 创建新的 IP 地址组
-          必填: name、businessType
-          常用: ipRanges、addressType、description、important
-          风险提示: 写操作；创建后可被防火墙策略引用
-          是否任务型: 否
-        - update_ipgroup
-          用途: 增量更新（PATCH）指定 IP 地址组
-          必填: uuid
-          常用: name、ipRanges、description
-          风险提示: 写操作；修改 IP 组会影响引用该组的所有策略
-          是否任务型: 否
-        - delete_ipgroup
-          用途: 删除指定 IP 地址组
-          必填: uuid
-          风险提示: 高风险写操作；如有策略引用该组将删除失败
-          是否任务型: 否
-
-        ## 服务对象
-        - get_services
-          用途: 查询服务或服务组列表（预定义或自定义）
-          必填: 无
-          常用: _start、_length、_search、serviceType
-          风险提示: 只读查询接口
-          是否任务型: 否
-        - get_service
-          用途: 获取单个服务或服务组详情
-          必填: uuid
-          风险提示: 只读查询接口
-          是否任务型: 否
-      enum:
-        - get_ipgroups
-        - get_ipgroup
-        - create_ipgroup
-        - update_ipgroup
-        - delete_ipgroup
-        - get_services
-        - get_service
-
-    uuid:
-      type: string
-      description: IP地址组或服务对象的唯一标识符（32字符UUID）
-    name:
-      type: string
-      description: 对象名称（最大95字符）
-    businessType:
-      type: string
-      description: >
-        IP地址组业务类型：IP=IP地址，ADDRGROUP=地址组，
-        USER=用户地址，BUSINESS=业务地址
-      enum:
-        - IP
-        - ADDRGROUP
-        - USER
-        - BUSINESS
-    addressType:
-      type: string
-      description: "IP协议版本：IPV4 或 IPV6"
-      enum:
-        - IPV4
-        - IPV6
-    important:
-      type: string
-      description: "重要级别：COMMON=普通，CORE=核心"
-      enum:
-        - COMMON
-        - CORE
-    ipRanges:
-      type: array
-      items:
-        type: object
-        properties:
-          start:
-            type: string
-            description: IP范围起始地址（如 192.168.1.1）
-          end:
-            type: string
-            description: IP范围结束地址（如 192.168.1.254）
-      description: IP地址范围列表
-    description:
-      type: string
-      description: 对象描述（最大95字符）
-    creator:
-      type: string
-      description: 创建者名称
-    serviceType:
-      type: string
-      description: "服务类型过滤：SERVICE=单个服务，SERVICEGROUP=服务组"
-      enum:
-        - SERVICE
-        - SERVICEGROUP
-
-    # Pagination
-    _start:
-      type: integer
-      description: 分页起始位置（从0开始）
-    _length:
-      type: integer
-      description: 每页最大返回数量（最大200，默认100）
-    __nameprefix:
-      type: string
-      description: 按名称前缀过滤（最大95字符）
-    _search:
-      type: string
-      description: 模糊搜索关键字（最大95字符）
-    _order:
-      type: string
-      description: "排序方向：asc 或 desc"
-      enum:
-        - asc
-        - desc
-    _sortby:
-      type: string
-      description: 排序字段名
-  required:
-    - action
-handler:
-  type: script
-  script_file: af.handler.py
-  function: objects
