--- .flocks/plugins/tools/api/af_v8_0_85/_provider.yaml
@@ -1,56 +1,0 @@
-name: af
-service_id: af_api_v8_0_85
-version: "8.0.85"
-description: >
-  Sangfor AF (Application Firewall) v8.0.85 REST API service.
-  Uses session-based cookie authentication: call login to obtain
-  a token, then supply it via Cookie header in subsequent requests.
-  This version adds monitoring (session/traffic/statistics) APIs
-  compared to v8.0.48.
-description_cn: >
-  深信服 AF 下一代防火墙 v8.0.85 REST API 服务（对应 AF8.0.95 发布版文档）。
-  使用基于 Session 的 Cookie 认证：首先调用 login 获取 token，
-  后续请求在 Cookie 中携带 token。在 v8.0.48 基础上新增监控（会话/流量/统计）相关 API。
-auth:
-  type: custom
-  secret: af_v8_0_85_username
-  secret_secret: af_v8_0_85_password
-credential_fields:
-  - key: username
-    label: 管理员用户名
-    storage: secret
-    config_key: username
-    secret_id: af_v8_0_85_username
-    input_type: text
-    required: true
-  - key: password
-    label: 管理员密码
-    storage: secret
-    config_key: password
-    secret_id: af_v8_0_85_password
-    input_type: password
-    required: true
-  - key: base_url
-    label: 设备地址 (Base URL)
-    storage: config
-    config_key: base_url
-    input_type: url
-    default: "https://192.168.1.1"
-  - key: verify_ssl
-    label: 验证 SSL 证书
-    storage: config
-    config_key: verify_ssl
-    input_type: text
-    default: "false"
-defaults:
-  base_url: "https://192.168.1.1"
-  timeout: 60
-  category: custom
-  product_version: "8.0.85"
-notes: |
-  深信服 AF v8.0.85 API 认证流程（同 v8.0.48）：
-  1. POST /api/v1/namespaces/public/login，传入用户名/密码，获取 token。
-  2. 后续请求在 HTTP 请求头中携带 Cookie: token=<token>。
-  3. token 默认 10 分钟无操作后失效，可通过 keepalive 接口刷新。
-  4. verify_ssl 默认 false（设备通常使用自签名证书）。
-  5. 本版本在 v8.0.48 基础上新增了监控相关 API（会话、流量统计等）。
