--- .flocks/plugins/tools/api/af_v8_0_85/af_auth.yaml
@@ -1,44 +1,0 @@
-name: af_auth
-description: >
-  Sangfor AF v8.0.48 authentication tool. Use the `action` parameter to
-  login, logout, or keep the session alive. Token is cached automatically
-  after a successful login.
-description_cn: >
-  深信服 AF v8.0.48 认证工具。通过 `action` 参数调用登录、注销或 token 保活接口。
-  登录成功后 token 会自动缓存，后续调用无需手动传 token。
-category: custom
-enabled: true
-requires_confirmation: false
-provider: af_api_v8_0_85
-inputSchema:
-  type: object
-  properties:
-    action:
-      type: string
-      description: |
-        认证动作名，可选值：
-        - login
-          用途: 登录设备，获取 session token（token 自动缓存）
-          必填: 无（用户名/密码从服务配置读取）
-          风险提示: 只读认证接口
-          是否任务型: 否
-        - logout
-          用途: 注销当前登录 session，清除 token 缓存
-          必填: 无
-          风险提示: 写操作，注销后需重新登录
-          是否任务型: 否
-        - keepalive
-          用途: 刷新 token 超时计时器，保持 session 活跃
-          必填: 无
-          风险提示: 只读接口
-          是否任务型: 否
-      enum:
-        - login
-        - logout
-        - keepalive
-  required:
-    - action
-handler:
-  type: script
-  script_file: af.handler.py
-  function: auth
