+++ .flocks/plugins/tools/api/sangfor_af_v8_0_106/_provider.yaml
@@ -1,0 +1,52 @@
+name: sangfor_af
+service_id: sangfor_af
+version: "8.0.106"
+description: >
+  Sangfor AF (Application Firewall) v8.0.106 REST API service (latest).
+  Includes all features from v8.0.85 plus extended session management
+  (recheck, export, block), alarm notifications, CPU/memory trend curves,
+  and link probe (SLA) APIs.
+description_cn: >
+  深信服 AF 下一代防火墙 v8.0.106 REST API 服务（最新版本）。
+  在 v8.0.85 基础上新增：会话精细化管理（recheck/导出/阻断）、
+  告警通知（在册事件、服务商信息、信息模板）、CPU/内存曲线、
+  链路探测（SLA）等接口。
+auth:
+  type: custom
+  secret: sangfor_af_v8_0_106_username
+  secret_secret: sangfor_af_v8_0_106_password
+credential_fields:
+  - key: username
+    label: 管理员用户名
+    storage: secret
+    config_key: username
+    secret_id: sangfor_af_v8_0_106_username
+    input_type: text
+    required: true
+  - key: password
+    label: 管理员密码
+    storage: secret
+    config_key: password
+    secret_id: sangfor_af_v8_0_106_password
+    input_type: password
+    required: true
+  - key: base_url
+    label: 设备地址 (Base URL)
+    storage: config
+    config_key: base_url
+    input_type: url
+    default: "https://192.168.1.1"
+defaults:
+  base_url: "https://192.168.1.1"
+  timeout: 60
+  category: custom
+  product_version: "8.0.106"
+notes: |
+  深信服 AF v8.0.106 API 认证流程（同 v8.0.85）：
+  1. 在 AF WebUI「系统 → 管理员账号」勾选 WEBAPI 权限。
+  2. Handler 自动用用户名/密码换取 token 并缓存（带 keepalive 自动续期）。
+  3. 后续所有请求由 Handler 自动注入 Cookie: token=<token>。
+  4. token 默认 10 分钟无操作后失效，缓存失效会自动重新登录。
+  5. 新增：会话清除/阻断、告警通知、CPU/内存趋势曲线、链路探测等 API。
+
+  `verify_ssl` 由表单底部「SSL 验证」开关控制（默认关闭，与 sangfor_sip 一致）。
