+++ .flocks/plugins/tools/api/sangfor_af_v8_0_106/_test.yaml
@@ -1,0 +1,144 @@
+schema_version: 1
+provider: sangfor_af
+
+connectivity:
+  tool: sangfor_af_v106_status
+  params:
+    action: get_system_version
+
+fixtures:
+  sangfor_af_v106_auth:
+    - label: "Session keepalive"
+      label_cn: "刷新 Session 保活"
+      tags: [smoke]
+      params:
+        action: keepalive
+      assert:
+        success: true
+
+  sangfor_af_v106_status:
+    - label: "Get system version"
+      label_cn: "获取系统版本信息"
+      tags: [smoke]
+      params:
+        action: get_system_version
+      assert:
+        success: true
+
+    - label: "Get CPU usage"
+      label_cn: "获取 CPU 使用率"
+      tags: [smoke]
+      params:
+        action: get_cpu_usage
+
+    - label: "Get CPU usage trend (last 60 min)"
+      label_cn: "获取近60分钟 CPU 使用率趋势"
+      tags: [smoke, trend]
+      params:
+        action: get_cpu_trend
+        minutes: 60
+
+    - label: "Get memory usage trend (last 60 min)"
+      label_cn: "获取近60分钟内存使用率趋势"
+      tags: [trend]
+      params:
+        action: get_memory_trend
+        minutes: 60
+
+  sangfor_af_v106_monitor:
+    - label: "Get session summary"
+      label_cn: "获取会话概要信息"
+      tags: [smoke, monitor]
+      params:
+        action: get_session_summary
+      assert:
+        success: true
+
+    - label: "Get daily new sessions"
+      label_cn: "获取每日新建会话信息"
+      tags: [monitor]
+      params:
+        action: get_session_dailys
+
+    - label: "Get user traffic top 10"
+      label_cn: "获取用户流量排行前10名"
+      tags: [monitor, traffic]
+      params:
+        action: get_user_traffic_rank
+        topNumber: 10
+
+    - label: "Get active sessions"
+      label_cn: "获取实时活跃会话列表"
+      tags: [monitor, session]
+      params:
+        action: get_sessions
+
+    - label: "Get session recheck config"
+      label_cn: "获取 Session Recheck 配置"
+      tags: [monitor, config]
+      params:
+        action: get_session_recheck
+
+  sangfor_af_v106_alarm:
+    - label: "Get alarm events config"
+      label_cn: "获取告警事件配置"
+      tags: [smoke, alarm]
+      params:
+        action: get_alarm_events_config
+      assert:
+        success: true
+
+    - label: "Get alarm notifications config"
+      label_cn: "获取告警通知渠道配置"
+      tags: [alarm]
+      params:
+        action: get_alarm_notifications
+
+    - label: "Get alarm messages (recent)"
+      label_cn: "获取历史告警消息列表"
+      tags: [alarm]
+      params:
+        action: get_alarm_messages
+        _length: 20
+        _order: desc
+
+  sangfor_af_v106_ops:
+    - label: "List blocked attacker IPs"
+      label_cn: "查询封锁攻击者 IP 列表"
+      tags: [smoke, blockip]
+      params:
+        action: get_blockip_list
+      assert:
+        success: true
+
+    - label: "List blacklist entries"
+      label_cn: "查询黑名单列表"
+      tags: [smoke, blacklist]
+      params:
+        action: get_blackwhitelist
+        type: BLACK
+
+  sangfor_af_v106_objects:
+    - label: "List IP groups"
+      label_cn: "查询 IP 地址组列表"
+      tags: [smoke]
+      params:
+        action: get_ipgroups
+      assert:
+        success: true
+
+    - label: "List link probes (SLA objects)"
+      label_cn: "查询链路探测对象列表"
+      tags: [smoke, sla]
+      params:
+        action: get_link_probes
+
+  sangfor_af_v106_network:
+    - label: "List routing table (all routes)"
+      label_cn: "查询完整路由表"
+      tags: [smoke]
+      params:
+        action: get_routes
+        routeType: ALL_ROUTE
+      assert:
+        success: true
