+++ .flocks/plugins/tools/api/sangfor_af_v8_0_106/sangfor_af_v106_alarm.yaml
@@ -1,0 +1,70 @@
+name: sangfor_af_v106_alarm
+description: >
+  Sangfor AF v8.0.106 alarm and notification tool. Query alarm event
+  configurations, notification channels, registered events, provider
+  information, message templates, and alarm message logs.
+  These APIs are available from v8.0.106 onwards.
+description_cn: >
+  深信服 AF v8.0.106 告警通知工具（v8.0.85 及以下不含此功能）。通过 `action` 参数
+  查询告警事件配置、告警通知渠道、在册事件、服务商信息、信息模板及告警消息日志。
+category: custom
+enabled: true
+requires_confirmation: false
+provider: sangfor_af
+inputSchema:
+  type: object
+  properties:
+    action:
+      type: string
+      description: |
+        告警通知动作名，可选值：
+        - get_alarm_events_config
+          用途: 获取告警事件配置（触发条件、阈值等）
+          必填: 无
+          风险提示: 只读接口
+          是否任务型: 否
+        - get_alarm_notifications
+          用途: 获取告警通知渠道配置（邮件/短信/企业微信等）
+          必填: 无
+          风险提示: 只读接口
+          是否任务型: 否
+        - get_alarm_messages
+          用途: 获取历史告警消息列表（已触发的告警）
+          必填: 无
+          常用: _start、_length、_search、_order
+          风险提示: 只读接口
+          是否任务型: 否
+        - get_topn_config
+          用途: 获取 TOP N 监控配置（流量排行展示数量等）
+          必填: 无
+          风险提示: 只读接口
+          是否任务型: 否
+      enum:
+        - get_alarm_events_config
+        - get_alarm_notifications
+        - get_alarm_messages
+        - get_topn_config
+    _start:
+      type: integer
+      description: 分页起始位置（从0开始）
+    _length:
+      type: integer
+      description: 每页最大返回数量
+    _search:
+      type: string
+      description: 模糊搜索关键字
+    _order:
+      type: string
+      description: "排序方向：asc 或 desc"
+      enum:
+        - asc
+        - desc
+    _sortby:
+      type: string
+      description: 排序字段名
+  required:
+    - action
+handler:
+  type: script
+  script_file: sangfor_af.handler.py
+  function: alarm
