+++ .flocks/plugins/tools/api/sangfor_af_v8_0_48/_provider.yaml
@@ -1,0 +1,55 @@
+name: sangfor_af
+service_id: sangfor_af
+version: "8.0.48"
+description: >
+  Sangfor AF (Application Firewall) v8.0.48 REST API service.
+  Uses session-based cookie authentication: call login to obtain
+  a token, then supply it via Cookie header in subsequent requests.
+description_cn: >
+  深信服 AF 下一代防火墙 v8.0.48 REST API 服务。
+  使用基于 Session 的 Cookie 认证：首先调用 login 获取 token，
+  后续请求在 Cookie 中携带 token。基础 URL 为设备管理地址（如 https://192.168.1.1）。
+auth:
+  type: custom
+  secret: sangfor_af_v8_0_48_username
+  secret_secret: sangfor_af_v8_0_48_password
+credential_fields:
+  - key: username
+    label: 管理员用户名
+    storage: secret
+    config_key: username
+    secret_id: sangfor_af_v8_0_48_username
+    input_type: text
+    required: true
+  - key: password
+    label: 管理员密码
+    storage: secret
+    config_key: password
+    secret_id: sangfor_af_v8_0_48_password
+    input_type: password
+    required: true
+  - key: base_url
+    label: 设备地址 (Base URL)
+    storage: config
+    config_key: base_url
+    input_type: url
+    default: "https://192.168.1.1"
+defaults:
+  base_url: "https://192.168.1.1"
+  timeout: 60
+  category: custom
+  product_version: "8.0.48"
+notes: |
+  深信服 AF v8.0.48 API 认证流程：
+  1. 在 AF WebUI「系统 → 管理员账号」勾选 WEBAPI 权限。
+  2. Handler 自动调用 POST /api/v1/namespaces/public/login，
+     用配置的用户名/密码换取 token 并缓存（带 keepalive 自动续期）。
+  3. 后续所有请求由 Handler 自动注入 Cookie: token=<token>。
+  4. token 默认 10 分钟无操作后失效，缓存命中失败时会自动重新登录。
+
+  `verify_ssl` 由表单底部「SSL 验证」开关控制，下列字段都会被
+  识别为 SSL 验证开关，按以下优先级取值：
+    1. `verify_ssl`                  （主键）
+    2. `ssl_verify`                  （兼容别名）
+    3. `custom_settings.verify_ssl`  （WebUI 表单写入位置）
+    4. 兜底默认值 `False`（**默认关闭证书验证**，AF 设备通常使用自签名证书）
