+++ .flocks/plugins/tools/api/sangfor_af_v8_0_85/_provider.yaml
@@ -1,0 +1,52 @@
+name: sangfor_af
+service_id: sangfor_af
+version: "8.0.85"
+description: >
+  Sangfor AF (Application Firewall) v8.0.85 REST API service.
+  Uses session-based cookie authentication: call login to obtain
+  a token, then supply it via Cookie header in subsequent requests.
+  This version adds monitoring (session/traffic/statistics) APIs
+  compared to v8.0.48.
+description_cn: >
+  深信服 AF 下一代防火墙 v8.0.85 REST API 服务（对应 AF8.0.95 发布版文档）。
+  使用基于 Session 的 Cookie 认证：首先调用 login 获取 token，
+  后续请求在 Cookie 中携带 token。在 v8.0.48 基础上新增监控（会话/流量/统计）相关 API。
+auth:
+  type: custom
+  secret: sangfor_af_v8_0_85_username
+  secret_secret: sangfor_af_v8_0_85_password
+credential_fields:
+  - key: username
+    label: 管理员用户名
+    storage: secret
+    config_key: username
+    secret_id: sangfor_af_v8_0_85_username
+    input_type: text
+    required: true
+  - key: password
+    label: 管理员密码
+    storage: secret
+    config_key: password
+    secret_id: sangfor_af_v8_0_85_password
+    input_type: password
+    required: true
+  - key: base_url
+    label: 设备地址 (Base URL)
+    storage: config
+    config_key: base_url
+    input_type: url
+    default: "https://192.168.1.1"
+defaults:
+  base_url: "https://192.168.1.1"
+  timeout: 60
+  category: custom
+  product_version: "8.0.85"
+notes: |
+  深信服 AF v8.0.85 API 认证流程（同 v8.0.48）：
+  1. 在 AF WebUI「系统 → 管理员账号」勾选 WEBAPI 权限。
+  2. Handler 自动用用户名/密码换取 token 并缓存（带 keepalive 自动续期）。
+  3. 后续所有请求由 Handler 自动注入 Cookie: token=<token>。
+  4. token 默认 10 分钟无操作后失效，缓存失效会自动重新登录。
+  5. 本版本在 v8.0.48 基础上新增了监控相关 API（会话/流量统计等）。
+
+  `verify_ssl` 由表单底部「SSL 验证」开关控制（默认关闭，与 sangfor_sip 一致）。
