+++ .flocks/plugins/tools/api/sangfor_af_v8_0_85/_test.yaml
@@ -1,0 +1,100 @@
+schema_version: 1
+provider: sangfor_af
+
+connectivity:
+  tool: sangfor_af_v85_status
+  params:
+    action: get_system_version
+
+fixtures:
+  sangfor_af_v85_auth:
+    - label: "Session keepalive"
+      label_cn: "刷新 Session 保活"
+      tags: [smoke]
+      params:
+        action: keepalive
+      assert:
+        success: true
+
+  sangfor_af_v85_status:
+    - label: "Get system version"
+      label_cn: "获取系统版本信息"
+      tags: [smoke]
+      params:
+        action: get_system_version
+      assert:
+        success: true
+
+    - label: "Get CPU usage"
+      label_cn: "获取 CPU 使用率"
+      tags: [smoke]
+      params:
+        action: get_cpu_usage
+
+  sangfor_af_v85_monitor:
+    - label: "Get session summary"
+      label_cn: "获取会话概要信息"
+      tags: [smoke, monitor]
+      params:
+        action: get_session_summary
+      assert:
+        success: true
+
+    - label: "Get daily new sessions"
+      label_cn: "获取每日新建会话信息"
+      tags: [monitor]
+      params:
+        action: get_session_dailys
+
+    - label: "Get user traffic top 10"
+      label_cn: "获取用户流量排行前10名"
+      tags: [monitor, traffic]
+      params:
+        action: get_user_traffic_rank
+        topNumber: 10
+
+    - label: "Get app traffic ranking"
+      label_cn: "获取应用流量排行"
+      tags: [monitor, traffic]
+      params:
+        action: get_app_traffic_rank
+
+    - label: "Get active sessions"
+      label_cn: "获取实时活跃会话列表"
+      tags: [monitor, session]
+      params:
+        action: get_sessions
+
+  sangfor_af_v85_ops:
+    - label: "List blocked attacker IPs"
+      label_cn: "查询封锁攻击者 IP 列表"
+      tags: [smoke, blockip]
+      params:
+        action: get_blockip_list
+      assert:
+        success: true
+
+    - label: "List blacklist entries"
+      label_cn: "查询黑名单列表"
+      tags: [smoke, blacklist]
+      params:
+        action: get_blackwhitelist
+        type: BLACK
+
+  sangfor_af_v85_objects:
+    - label: "List IP groups"
+      label_cn: "查询 IP 地址组列表"
+      tags: [smoke]
+      params:
+        action: get_ipgroups
+      assert:
+        success: true
+
+  sangfor_af_v85_network:
+    - label: "List routing table"
+      label_cn: "查询路由表"
+      tags: [smoke]
+      params:
+        action: get_routes
+      assert:
+        success: true
